minor: make diagnostics more similar
